@@ -12,107 +12,104 @@
 //
 //===----------------------------------------------------------------------===//
 
+import Foundation
 import MetricsTestKit
-import XCTest
+import Testing
 
 @testable import CoreMetrics
 @testable import Metrics
 
-class MetricsExtensionsTests: XCTestCase {
-    func testTimerBlock() throws {
-        // bootstrap with our test metrics
+#if canImport(Dispatch)
+import Dispatch
+#endif
+
+struct MetricsExtensionsTests {
+    #if canImport(Dispatch)
+    @Test func timerBlock() throws {
+        // create our test metrics, avoid bootstrapping global MetricsSystem
         let metrics = TestMetrics()
-        MetricsSystem.bootstrapInternal(metrics)
         // run the test
         let name = "timer-\(UUID().uuidString)"
         let delay = 0.05
-        Timer.measure(label: name) {
+        Timer.measure(label: name, factory: metrics) {
             Thread.sleep(forTimeInterval: delay)
         }
         let timer = try metrics.expectTimer(name)
-        XCTAssertEqual(1, timer.values.count, "expected number of entries to match")
-        XCTAssertGreaterThan(timer.values[0], Int64(delay * 1_000_000_000), "expected delay to match")
+        #expect(timer.values.count == 1, "expected number of entries to match")
+        #expect(timer.values[0] > Int64(delay * 1_000_000_000), "expected delay to match")
     }
+    #endif
 
-    func testTimerWithTimeInterval() throws {
-        // bootstrap with our test metrics
+    @Test func timerWithTimeInterval() throws {
+        // create our test metrics, avoid bootstrapping global MetricsSystem
         let metrics = TestMetrics()
-        MetricsSystem.bootstrapInternal(metrics)
         // run the test
-        let timer = Timer(label: "test-timer")
+        let timer = Timer(label: "test-timer", factory: metrics)
         let testTimer = try metrics.expectTimer(timer)
         let timeInterval = TimeInterval(Double.random(in: 1...500))
         timer.record(timeInterval)
-        XCTAssertEqual(1, testTimer.values.count, "expected number of entries to match")
-        XCTAssertEqual(testTimer.values[0], Int64(timeInterval * 1_000_000_000), "expected value to match")
+        #expect(testTimer.values.count == 1, "expected number of entries to match")
+        #expect(testTimer.values[0] == Int64(timeInterval * 1_000_000_000), "expected value to match")
     }
 
-    func testTimerWithDispatchTime() throws {
-        // bootstrap with our test metrics
+    #if canImport(Dispatch)
+    @Test func timerWithDispatchTime() throws {
+        // create our test metrics, avoid bootstrapping global MetricsSystem
         let metrics = TestMetrics()
-        MetricsSystem.bootstrapInternal(metrics)
         // run the test
-        let timer = Timer(label: "test-timer")
+        let timer = Timer(label: "test-timer", factory: metrics)
         let testTimer = try metrics.expectTimer(timer)
         // nano
         let nano = DispatchTimeInterval.nanoseconds(Int.random(in: 1...500))
         timer.record(nano)
-        XCTAssertEqual(testTimer.values.count, 1, "expected number of entries to match")
-        XCTAssertEqual(Int(testTimer.values[0]), nano.nano(), "expected value to match")
+        #expect(testTimer.values.count == 1, "expected number of entries to match")
+        #expect(Int(testTimer.values[0]) == nano.nano(), "expected value to match")
         // micro
         let micro = DispatchTimeInterval.microseconds(Int.random(in: 1...500))
         timer.record(micro)
-        XCTAssertEqual(testTimer.values.count, 2, "expected number of entries to match")
-        XCTAssertEqual(Int(testTimer.values[1]), micro.nano(), "expected value to match")
+        #expect(testTimer.values.count == 2, "expected number of entries to match")
+        #expect(Int(testTimer.values[1]) == micro.nano(), "expected value to match")
         // milli
         let milli = DispatchTimeInterval.milliseconds(Int.random(in: 1...500))
         timer.record(milli)
-        XCTAssertEqual(testTimer.values.count, 3, "expected number of entries to match")
-        XCTAssertEqual(Int(testTimer.values[2]), milli.nano(), "expected value to match")
+        #expect(testTimer.values.count == 3, "expected number of entries to match")
+        #expect(Int(testTimer.values[2]) == milli.nano(), "expected value to match")
         // seconds
         let sec = DispatchTimeInterval.seconds(Int.random(in: 1...500))
         timer.record(sec)
-        XCTAssertEqual(testTimer.values.count, 4, "expected number of entries to match")
-        XCTAssertEqual(Int(testTimer.values[3]), sec.nano(), "expected value to match")
+        #expect(testTimer.values.count == 4, "expected number of entries to match")
+        #expect(Int(testTimer.values[3]) == sec.nano(), "expected value to match")
         // never
         timer.record(DispatchTimeInterval.never)
-        XCTAssertEqual(testTimer.values.count, 5, "expected number of entries to match")
-        XCTAssertEqual(testTimer.values[4], 0, "expected value to match")
+        #expect(testTimer.values.count == 5, "expected number of entries to match")
+        #expect(testTimer.values[4] == 0, "expected value to match")
     }
 
-    func testTimerWithDispatchTimeInterval() throws {
+    @Test func timerWithDispatchTimeInterval() throws {
         let metrics = TestMetrics()
-        MetricsSystem.bootstrapInternal(metrics)
 
         let name = "timer-\(UUID().uuidString)"
 
-        let timer = Timer(label: name)
+        let timer = Timer(label: name, factory: metrics)
         let start = DispatchTime.now()
         let end = DispatchTime(uptimeNanoseconds: start.uptimeNanoseconds + 1000 * 1000 * 1000)
         timer.recordInterval(since: start, end: end)
 
         let testTimer = try metrics.expectTimer(timer)
-        XCTAssertEqual(testTimer.values.count, 1, "expected number of entries to match")
-        XCTAssertEqual(
-            UInt64(testTimer.values.first!),
-            end.uptimeNanoseconds - start.uptimeNanoseconds,
+        #expect(testTimer.values.count == 1, "expected number of entries to match")
+        #expect(
+            UInt64(testTimer.values.first!) == end.uptimeNanoseconds - start.uptimeNanoseconds,
             "expected value to match"
         )
-        XCTAssertEqual(metrics.timers.count, 1, "timer should have been stored")
+        #expect(metrics.timers.count == 1, "timer should have been stored")
     }
+    #endif
 
-    func testTimerDuration() throws {
-        // Wrapping only the insides of the test case so that the generated
-        // tests on Linux in MetricsTests+XCTest don't complain that the func does not exist.
-        guard #available(iOS 16, macOS 13, tvOS 15, watchOS 8, *) else {
-            throw XCTSkip("Timer.record(_ duration: Duration) is not available on this platform")
-        }
-
+    @Test func timerDuration() throws {
         let metrics = TestMetrics()
-        MetricsSystem.bootstrapInternal(metrics)
 
         let name = "timer-\(UUID().uuidString)"
-        let timer = Timer(label: name)
+        let timer = Timer(label: name, factory: metrics)
 
         let duration = Duration(secondsComponent: 3, attosecondsComponent: 123_000_000_000_000_000)
         let nanoseconds = duration.components.seconds * 1_000_000_000 + duration.components.attoseconds / 1_000_000_000
@@ -123,105 +120,144 @@ class MetricsExtensionsTests: XCTestCase {
         timer.record(duration: Duration(secondsComponent: 10_000_000_000, attosecondsComponent: 123))
 
         let testTimer = try metrics.expectTimer(timer)
-        XCTAssertEqual(testTimer.values.count, 2, "expected number of entries to match")
-        XCTAssertEqual(testTimer.values.first, nanoseconds, "expected value to match")
-        XCTAssertEqual(testTimer.values[1], Int64.max, "expected to record Int64.max if Durataion overflows")
-        XCTAssertEqual(metrics.timers.count, 1, "timer should have been stored")
+        #expect(testTimer.values.count == 2, "expected number of entries to match")
+        #expect(testTimer.values.first == nanoseconds, "expected value to match")
+        #expect(testTimer.values[1] == Int64.max, "expected to record Int64.max if Durataion overflows")
+        #expect(metrics.timers.count == 1, "timer should have been stored")
     }
 
-    func testTimerUnits() throws {
+    @Test func timerUnits() throws {
         let metrics = TestMetrics()
-        MetricsSystem.bootstrapInternal(metrics)
 
         let name = "timer-\(UUID().uuidString)"
         let value = Int64.random(in: 0...1000)
 
-        let timer = Timer(label: name)
+        let timer = Timer(label: name, factory: metrics)
         timer.recordNanoseconds(value)
 
         let testTimer = try metrics.expectTimer(timer)
-        XCTAssertEqual(testTimer.values.count, 1, "expected number of entries to match")
-        XCTAssertEqual(testTimer.values.first, value, "expected value to match")
-        XCTAssertEqual(metrics.timers.count, 1, "timer should have been stored")
+        #expect(testTimer.values.count == 1, "expected number of entries to match")
+        #expect(testTimer.values.first == value, "expected value to match")
+        #expect(metrics.timers.count == 1, "timer should have been stored")
 
         let secondsName = "timer-seconds-\(UUID().uuidString)"
         let secondsValue = Int64.random(in: 0...1000)
-        let secondsTimer = Timer(label: secondsName, preferredDisplayUnit: .seconds)
+        let secondsTimer = Timer(label: secondsName, preferredDisplayUnit: .seconds, factory: metrics)
         secondsTimer.recordSeconds(secondsValue)
 
         let testSecondsTimer = try metrics.expectTimer(secondsTimer)
-        XCTAssertEqual(testSecondsTimer.values.count, 1, "expected number of entries to match")
-        XCTAssertEqual(metrics.timers.count, 2, "timer should have been stored")
+        #expect(testSecondsTimer.values.count == 1, "expected number of entries to match")
+        #expect(metrics.timers.count == 2, "timer should have been stored")
     }
 
-    func testPreferDisplayUnit() throws {
+    @Test func preferDisplayUnit() throws {
         let metrics = TestMetrics()
-        MetricsSystem.bootstrapInternal(metrics)
 
         let value = Double.random(in: 0...1000)
-        let timer = Timer(label: "test", preferredDisplayUnit: .seconds)
+        let timer = Timer(label: "test", preferredDisplayUnit: .seconds, factory: metrics)
         timer.recordSeconds(value)
 
         let testTimer = try metrics.expectTimer(timer)
 
+        // The suggested way for comparing float numbers is to use swift-numerics,
+        // but this would add a dependency. Instead, as we fully control the values
+        // used in tests, we can get away with the simple `abs(x - y) < accuracy`.
+        // See https://developer.apple.com/documentation/testing/migratingfromxctest
+
         testTimer.preferDisplayUnit(.nanoseconds)
-        XCTAssertEqual(
-            testTimer.valueInPreferredUnit(atIndex: 0),
-            value * 1000 * 1000 * 1000,
-            accuracy: 1.0,
+        #expect(
+            abs(testTimer.valueInPreferredUnit(atIndex: 0) - value * 1000 * 1000 * 1000) < 1.0,
             "expected value to match"
         )
 
         testTimer.preferDisplayUnit(.microseconds)
-        XCTAssertEqual(
-            testTimer.valueInPreferredUnit(atIndex: 0),
-            value * 1000 * 1000,
-            accuracy: 0.001,
+        #expect(
+            abs(testTimer.valueInPreferredUnit(atIndex: 0) - value * 1000 * 1000) < 0.1,
             "expected value to match"
         )
 
         testTimer.preferDisplayUnit(.milliseconds)
-        XCTAssertEqual(
-            testTimer.valueInPreferredUnit(atIndex: 0),
-            value * 1000,
-            accuracy: 0.000001,
+        #expect(
+            abs(testTimer.valueInPreferredUnit(atIndex: 0) - value * 1000) < 0.1,
             "expected value to match"
         )
 
         testTimer.preferDisplayUnit(.seconds)
-        XCTAssertEqual(
-            testTimer.valueInPreferredUnit(atIndex: 0),
-            value,
-            accuracy: 0.000000001,
+        #expect(
+            abs(testTimer.valueInPreferredUnit(atIndex: 0) - value) < 0.000000001,
             "expected value to match"
         )
 
         testTimer.preferDisplayUnit(.minutes)
-        XCTAssertEqual(
-            testTimer.valueInPreferredUnit(atIndex: 0),
-            value / 60,
-            accuracy: 0.000000001,
+        #expect(
+            abs(testTimer.valueInPreferredUnit(atIndex: 0) - value / 60) < 0.000000001,
             "expected value to match"
         )
 
         testTimer.preferDisplayUnit(.hours)
-        XCTAssertEqual(
-            testTimer.valueInPreferredUnit(atIndex: 0),
-            value / (60 * 60),
-            accuracy: 0.000000001,
+        #expect(
+            abs(testTimer.valueInPreferredUnit(atIndex: 0) - value / (60 * 60)) < 0.000000001,
             "expected value to match"
         )
 
         testTimer.preferDisplayUnit(.days)
-        XCTAssertEqual(
-            testTimer.valueInPreferredUnit(atIndex: 0),
-            value / (60 * 60 * 24),
-            accuracy: 0.000000001,
+        #expect(
+            abs(testTimer.valueInPreferredUnit(atIndex: 0) - value / (60 * 60 * 24)) < 0.000000001,
             "expected value to match"
         )
     }
+
+    @Test func timerMeasure() async throws {
+        // create our test metrics, avoid bootstrapping global MetricsSystem
+        let metrics = TestMetrics()
+        // run the test
+        let name = "timer-\(UUID().uuidString)"
+        let delay = Duration.milliseconds(5)
+        let timer = Timer(label: name, factory: metrics)
+        try await timer.measure {
+            try await Task.sleep(for: delay)
+        }
+
+        let expectedTimer = try metrics.expectTimer(name)
+        #expect(expectedTimer.values.count == 1, "expected number of entries to match")
+        #expect(expectedTimer.values[0] > delay.nanosecondsClamped, "expected delay to match")
+    }
+
+    @Test func timerMeasureSync() async throws {
+        // create our test metrics, avoid bootstrapping global MetricsSystem
+        let metrics = TestMetrics()
+        // run the test
+        let name = "timer-\(UUID().uuidString)"
+        let delay = 0.5
+        let timer = Timer(label: name, factory: metrics)
+        timer.measure {
+            Thread.sleep(forTimeInterval: delay)
+        }
+
+        let expectedTimer = try metrics.expectTimer(name)
+        #expect(expectedTimer.values.count == 1, "expected number of entries to match")
+        #expect(expectedTimer.values[0] > Int64(delay * 1_000_000_000), "expected delay to match in nanoseconds")
+    }
+
+    @MainActor
+    @Test func timerMeasureFromMainActor() async throws {
+        // create our test metrics, avoid bootstrapping global MetricsSystem
+        let metrics = TestMetrics()
+        // run the test
+        let name = "timer-\(UUID().uuidString)"
+        let delay = Duration.milliseconds(5)
+        let timer = Timer(label: name, factory: metrics)
+        try await timer.measure {
+            try await Task.sleep(for: delay)
+        }
+
+        let expectedTimer = try metrics.expectTimer(name)
+        #expect(expectedTimer.values.count == 1, "expected number of entries to match")
+        #expect(expectedTimer.values[0] > delay.nanosecondsClamped, "expected delay to match")
+    }
 }
 
+#if canImport(Dispatch)
 // https://bugs.swift.org/browse/SR-6310
 extension DispatchTimeInterval {
     func nano() -> Int {
@@ -249,5 +285,26 @@ extension DispatchTimeInterval {
         default:
             return 0
         }
+    }
+}
+#endif
+
+@available(macOS 13, iOS 16, tvOS 16, watchOS 9, *)
+extension Swift.Duration {
+    fileprivate var nanosecondsClamped: Int64 {
+        let components = self.components
+
+        let secondsComponentNanos = components.seconds.multipliedReportingOverflow(by: 1_000_000_000)
+        let attosCompononentNanos = components.attoseconds / 1_000_000_000
+        let combinedNanos = secondsComponentNanos.partialValue.addingReportingOverflow(attosCompononentNanos)
+
+        guard
+            !secondsComponentNanos.overflow,
+            !combinedNanos.overflow
+        else {
+            return .max
+        }
+
+        return combinedNanos.partialValue
     }
 }
