@@ -12,22 +12,21 @@
 //
 //===----------------------------------------------------------------------===//
 
-import Dispatch
+import Foundation
 import MetricsTestKit
-import XCTest
+import Testing
 
 @testable import CoreMetrics
 
-class SendableTest: XCTestCase {
-    func testSendableMetrics() async throws {
-        // bootstrap with our test metrics
+struct SendableTest {
+    @Test func sendableMetrics() async throws {
+        // create our test metrics, avoid bootstrapping global MetricsSystem
         let metrics = TestMetrics()
-        MetricsSystem.bootstrapInternal(metrics)
 
         do {
             let name = "counter-\(UUID().uuidString)"
             let value = Int.random(in: 0...1000)
-            let counter = Counter(label: name)
+            let counter = Counter(label: name, factory: metrics)
 
             let task = Task.detached { () -> [Int64] in
                 counter.increment(by: value)
@@ -35,14 +34,14 @@ class SendableTest: XCTestCase {
                 return handler.values
             }
             let values = try await task.value
-            XCTAssertEqual(values.count, 1, "expected number of entries to match")
-            XCTAssertEqual(values[0], Int64(value), "expected value to match")
+            #expect(values.count == 1, "expected number of entries to match")
+            #expect(values[0] == Int64(value), "expected value to match")
         }
 
         do {
             let name = "floating-point-counter-\(UUID().uuidString)"
             let value = Double.random(in: 0...0.9999)
-            let counter = FloatingPointCounter(label: name)
+            let counter = FloatingPointCounter(label: name, factory: metrics)
 
             let task = Task.detached { () -> Double in
                 counter.increment(by: value)
@@ -50,13 +49,13 @@ class SendableTest: XCTestCase {
                 return handler.fraction
             }
             let fraction = await task.value
-            XCTAssertEqual(fraction, value)
+            #expect(fraction == value)
         }
 
         do {
             let name = "recorder-\(UUID().uuidString)"
             let value = Double.random(in: -1000...1000)
-            let recorder = Recorder(label: name)
+            let recorder = Recorder(label: name, factory: metrics)
 
             let task = Task.detached { () -> [Double] in
                 recorder.record(value)
@@ -64,14 +63,14 @@ class SendableTest: XCTestCase {
                 return handler.values
             }
             let values = try await task.value
-            XCTAssertEqual(values.count, 1, "expected number of entries to match")
-            XCTAssertEqual(values[0], value, "expected value to match")
+            #expect(values.count == 1, "expected number of entries to match")
+            #expect(values[0] == value, "expected value to match")
         }
 
         do {
             let name = "meter-\(UUID().uuidString)"
             let value = Double.random(in: -1000...1000)
-            let meter = Meter(label: name)
+            let meter = Meter(label: name, factory: metrics)
 
             let task = Task.detached { () -> [Double] in
                 meter.set(value)
@@ -79,14 +78,14 @@ class SendableTest: XCTestCase {
                 return handler.values
             }
             let values = try await task.value
-            XCTAssertEqual(values.count, 1, "expected number of entries to match")
-            XCTAssertEqual(values[0], value, "expected value to match")
+            #expect(values.count == 1, "expected number of entries to match")
+            #expect(values[0] == value, "expected value to match")
         }
 
         do {
             let name = "timer-\(UUID().uuidString)"
             let value = Int64.random(in: 0...1000)
-            let timer = Timer(label: name)
+            let timer = Timer(label: name, factory: metrics)
 
             let task = Task.detached { () -> [Int64] in
                 timer.recordNanoseconds(value)
@@ -94,8 +93,8 @@ class SendableTest: XCTestCase {
                 return handler.values
             }
             let values = try await task.value
-            XCTAssertEqual(values.count, 1, "expected number of entries to match")
-            XCTAssertEqual(values[0], value, "expected value to match")
+            #expect(values.count == 1, "expected number of entries to match")
+            #expect(values[0] == value, "expected value to match")
         }
     }
 }
