@@ -17,38 +17,74 @@
 // MARK: - Counter
 
 /// A counter is a cumulative metric that represents a single monotonically increasing counter whose value can only increase or be reset to zero.
+///
 /// For example, you can use a counter to represent the number of requests served, tasks completed, or errors.
 ///
-/// This is the user-facing Counter API.
+/// This is the user-facing counter API.
+/// Its behavior is defined by the ``CounterHandler`` implementation.
 ///
-/// Its behavior depends on the `CounterHandler` implementation.
+/// Increment a counter:
+///
+/// ```swift
+/// counter.increment(by: 5)
+/// counter.increment() // increments counter by 1
+/// ```
+///
+/// Reset a counter:
+///
+/// ```swift
+/// counter.reset()
+/// ````
 public final class Counter {
-    /// ``_handler`` is only public to allow access from `MetricsTestKit`. Do not consider it part of the public API.
+    /// `_handler` and `_factory` are only public to allow access from `MetricsTestKit`.
+    /// Do not consider them part of the public API.
+    @_documentation(visibility: internal)
     public let _handler: CounterHandler
+    @_documentation(visibility: internal)
+    @usableFromInline
+    package let _factory: MetricsFactory
+    /// The label for the counter.
     public let label: String
+    /// The dimensions for the counter.
     public let dimensions: [(String, String)]
 
-    /// Alternative way to create a new `Counter`, while providing an explicit `CounterHandler`.
-    ///
-    /// - warning: This initializer provides an escape hatch for situations where one must use a custom factory instead of the global one.
-    ///            We do not expect this API to be used in normal circumstances, so if you find yourself using it make sure it's for a good reason.
+    /// Alternative way to create a new counter with an explicit counter handler.
     ///
     /// - SeeAlso: Use `init(label:dimensions:)` to create `Counter` instances using the configured metrics backend.
     ///
     /// - parameters:
-    ///     - label: The label for the `Counter`.
-    ///     - dimensions: The dimensions for the `Counter`.
-    ///     - handler: The custom backend.
-    public init(label: String, dimensions: [(String, String)], handler: CounterHandler) {
+    ///   - label: The label for the `Counter`.
+    ///   - dimensions: The dimensions for the `Counter`.
+    ///   - handler: The custom backend.
+    ///   - factory: The custom metrics factory.
+    public init(label: String, dimensions: [(String, String)], handler: CounterHandler, factory: MetricsFactory) {
         self.label = label
         self.dimensions = dimensions
         self._handler = handler
+        self._factory = factory
+    }
+
+    /// Alternative way to create a new counter, with an explicit counter handler.
+    ///
+    /// - SeeAlso: Use `init(label:dimensions:)` to create instances of ``Counter`` using the configured metrics backend.
+    ///
+    /// - parameters:
+    ///   - label: The label for the `Counter`.
+    ///   - dimensions: The dimensions for the `Counter`.
+    ///   - handler: The custom backend, created by the global metrics factory.
+    public convenience init(label: String, dimensions: [(String, String)], handler: CounterHandler) {
+        self.init(
+            label: label,
+            dimensions: dimensions,
+            handler: handler,
+            factory: MetricsSystem.factory
+        )
     }
 
     /// Increment the counter.
     ///
     /// - parameters:
-    ///     - amount: Amount to increment by.
+    ///   - amount: Amount to increment by.
     @inlinable
     public func increment<DataType: BinaryInteger>(by amount: DataType) {
         self._handler.increment(by: Int64(amount))
@@ -68,21 +104,32 @@ public final class Counter {
 }
 
 extension Counter {
-    /// Create a new `Counter`.
+    /// Create a new counter.
     ///
     /// - parameters:
-    ///     - label: The label for the `Counter`.
-    ///     - dimensions: The dimensions for the `Counter`.
+    ///   - label: The label for the `Counter`.
+    ///   - dimensions: The dimensions for the `Counter`.
     public convenience init(label: String, dimensions: [(String, String)] = []) {
-        let handler = MetricsSystem.factory.makeCounter(label: label, dimensions: dimensions)
-        self.init(label: label, dimensions: dimensions, handler: handler)
+        self.init(label: label, dimensions: dimensions, factory: MetricsSystem.factory)
+    }
+
+    /// Create a new counter using a custom metrics factory that you provide.
+    ///
+    /// - parameters:
+    ///   - label: The label for the `Counter`.
+    ///   - dimensions: The dimensions for the `Counter`.
+    ///   - factory: The custom metrics factory.
+    public convenience init(label: String, dimensions: [(String, String)] = [], factory: MetricsFactory) {
+        let handler = factory.makeCounter(label: label, dimensions: dimensions)
+        self.init(label: label, dimensions: dimensions, handler: handler, factory: factory)
     }
 
     /// Signal the underlying metrics library that this counter will never be updated again.
-    /// In response the library MAY decide to eagerly release any resources held by this `Counter`.
+    ///
+    /// In response the library MAY decide to eagerly release any resources held by this counter.
     @inlinable
     public func destroy() {
-        MetricsSystem.factory.destroyCounter(self._handler)
+        self._factory.destroyCounter(self._handler)
     }
 }
 
@@ -94,52 +141,93 @@ extension Counter: CustomStringConvertible {
 
 // MARK: - FloatingPointCounter
 
-/// A FloatingPointCounter is a cumulative metric that represents a single monotonically increasing FloatingPointCounter whose value can only increase or be reset to zero.
-/// For example, you can use a FloatingPointCounter to represent the number of requests served, tasks completed, or errors.
-/// FloatingPointCounter is not supported by all metrics backends, however a default implementation is provided which accumulates floating point values and records increments to a standard Counter after crossing integer boundaries.
+/// A floating-point counter is a cumulative metric that represents a single monotonically increasing floating-point counter whose value can only increase or be reset to zero.
 ///
-/// This is the user-facing FloatingPointCounter API.
+/// For example, you can use a floating-point counter to represent the number of requests served, tasks completed, or errors.
+/// floating-point counter is not supported by all metrics backends, however a default implementation is provided which accumulates floating point values and records increments to a standard counter after crossing integer boundaries.
 ///
-/// Its behavior depends on the `FloatingCounterHandler` implementation.
+/// This is the user-facing floating-point counter API.
+/// Its behavior depends on the ``FloatingPointCounterHandler`` implementation.
+///
+/// Increment a counter:
+///
+/// ```swift
+/// counter.increment(by: 5.1)
+/// counter.increment() // increments counter by 1
+/// ```
+///
+/// Reset a counter:
+///
+/// ```swift
+/// counter.reset()
+/// ````
 public final class FloatingPointCounter {
-    /// ``_handler`` is only public to allow access from `MetricsTestKit`. Do not consider it part of the public API.
+    /// `_handler` and `_factory` are only public to allow access from `MetricsTestKit`.
+    /// Do not consider them part of the public API.
+    @_documentation(visibility: internal)
     public let _handler: FloatingPointCounterHandler
+    @_documentation(visibility: internal)
+    @usableFromInline
+    package let _factory: MetricsFactory
+    /// The label for the floating point counter.
     public let label: String
+    /// The dimensions for the floating point counter.
     public let dimensions: [(String, String)]
 
-    /// Alternative way to create a new `FloatingPointCounter`, while providing an explicit `FloatingPointCounterHandler`.
+    /// Alternative way to create a new floating-point counter, while providing an explicit floating-point counter handler.
     ///
-    /// - warning: This initializer provides an escape hatch for situations where one must use a custom factory instead of the global one.
-    ///            We do not expect this API to be used in normal circumstances, so if you find yourself using it make sure it's for a good reason.
-    ///
-    /// - SeeAlso: Use `init(label:dimensions:)` to create `FloatingPointCounter` instances using the configured metrics backend.
+    /// - SeeAlso: Use `init(label:dimensions:)` to create instances of ``FloatingPointCounter``using the configured metrics backend.
     ///
     /// - parameters:
-    ///     - label: The label for the `FloatingPointCounter`.
-    ///     - dimensions: The dimensions for the `FloatingPointCounter`.
-    ///     - handler: The custom backend.
-    public init(label: String, dimensions: [(String, String)], handler: FloatingPointCounterHandler) {
+    ///  - label: The label for the `FloatingPointCounter`.
+    ///  - dimensions: The dimensions for the `FloatingPointCounter`.
+    ///  - handler: The custom backend.
+    ///  - factory: The custom metrics factory.
+    public init(
+        label: String,
+        dimensions: [(String, String)],
+        handler: FloatingPointCounterHandler,
+        factory: MetricsFactory
+    ) {
         self.label = label
         self.dimensions = dimensions
         self._handler = handler
+        self._factory = factory
     }
 
-    /// Increment the FloatingPointCounter.
+    /// Alternative way to create a new floating-point counter, while providing an explicit floating-point counter handler..
+    ///
+    /// - SeeAlso: Use `init(label:dimensions:)` to create instances of ``FloatingPointCounter`` using the configured metrics backend.
     ///
     /// - parameters:
-    ///     - amount: Amount to increment by.
+    ///   - label: The label for the `FloatingPointCounter`.
+    ///   - dimensions: The dimensions for the `FloatingPointCounter`.
+    ///   - handler: The custom backend.
+    public convenience init(label: String, dimensions: [(String, String)], handler: FloatingPointCounterHandler) {
+        self.init(
+            label: label,
+            dimensions: dimensions,
+            handler: handler,
+            factory: MetricsSystem.factory
+        )
+    }
+
+    /// Increment the floating-point counter.
+    ///
+    /// - parameters:
+    ///   - amount: Amount to increment by.
     @inlinable
     public func increment<DataType: BinaryFloatingPoint>(by amount: DataType) {
         self._handler.increment(by: Double(amount))
     }
 
-    /// Increment the FloatingPointCounter by one.
+    /// Increment the floating-point counter by one.
     @inlinable
     public func increment() {
         self.increment(by: 1.0)
     }
 
-    /// Reset the FloatingPointCounter back to zero.
+    /// Reset the floating-point counter back to zero.
     @inlinable
     public func reset() {
         self._handler.reset()
@@ -147,21 +235,32 @@ public final class FloatingPointCounter {
 }
 
 extension FloatingPointCounter {
-    /// Create a new `FloatingPointCounter`.
+    /// Create a new floating-point counter.
     ///
     /// - parameters:
-    ///     - label: The label for the `FloatingPointCounter`.
-    ///     - dimensions: The dimensions for the `FloatingPointCounter`.
+    ///   - label: The label for the `FloatingPointCounter`.
+    ///   - dimensions: The dimensions for the `FloatingPointCounter`.
     public convenience init(label: String, dimensions: [(String, String)] = []) {
-        let handler = MetricsSystem.factory.makeFloatingPointCounter(label: label, dimensions: dimensions)
-        self.init(label: label, dimensions: dimensions, handler: handler)
+        self.init(label: label, dimensions: dimensions, factory: MetricsSystem.factory)
     }
 
-    /// Signal the underlying metrics library that this FloatingPointCounter will never be updated again.
+    /// Create a new floating-point counter using a custom metrics factory that you provide.
+    ///
+    /// - parameters:
+    ///   - label: The label for the `FloatingPointCounter`.
+    ///   - dimensions: The dimensions for the `FloatingPointCounter`.
+    ///   - factory: The custom metrics factory.
+    public convenience init(label: String, dimensions: [(String, String)] = [], factory: MetricsFactory) {
+        let handler = factory.makeFloatingPointCounter(label: label, dimensions: dimensions)
+        self.init(label: label, dimensions: dimensions, handler: handler, factory: factory)
+    }
+
+    /// Signal the underlying metrics library that this floating point counter will never be updated again.
+    ///
     /// In response the library MAY decide to eagerly release any resources held by this `FloatingPointCounter`.
     @inlinable
     public func destroy() {
-        MetricsSystem.factory.destroyFloatingPointCounter(self._handler)
+        self._factory.destroyFloatingPointCounter(self._handler)
     }
 }
 
@@ -174,89 +273,130 @@ extension FloatingPointCounter: CustomStringConvertible {
 // MARK: - Gauge
 
 /// A gauge is a metric that represents a single numerical value that can arbitrarily go up and down.
-/// Gauges are typically used for measured values like temperatures or current memory usage, but also "counts" that can go up and down, like the number of active threads.
+///
+/// Gauges are typically used for measured values like temperatures or current memory usage, but also "counts" that can go up and down, such as the number of active threads.
 /// Gauges are modeled as `Recorder` with a sample size of 1 and that does not perform any aggregation.
+///
+/// Recording a value with a gauge:
+///
+/// ```swift
+/// guage.record(100)
+/// ```
 public final class Gauge: Recorder, @unchecked Sendable {
-    /// Create a new `Gauge`.
+    /// Create a new gauge.
     ///
     /// - parameters:
-    ///     - label: The label for the `Gauge`.
-    ///     - dimensions: The dimensions for the `Gauge`.
+    ///   - label: The label for the `Gauge`.
+    ///   - dimensions: The dimensions for the `Gauge`.
     public convenience init(label: String, dimensions: [(String, String)] = []) {
         self.init(label: label, dimensions: dimensions, aggregate: false)
+    }
+
+    /// Create a new gauge using a custom metrics factory that you provide.
+    ///
+    /// - parameters:
+    ///   - label: The label for the `Gauge`.
+    ///   - dimensions: The dimensions for the `Gauge`.
+    ///   - factory: The custom metrics factory.
+    public convenience init(label: String, dimensions: [(String, String)] = [], factory: MetricsFactory) {
+        self.init(label: label, dimensions: dimensions, aggregate: false, factory: factory)
     }
 }
 
 // MARK: - Meter
 
 /// A meter is similar to a gauge, it is a metric that represents a single numerical value that can arbitrarily go up and down.
-/// Meters are typically used for measured values like temperatures or current memory usage, but also "counts" that can go up and down, like the number of active threads.
+///
+/// Meters are typically used for measured values like temperatures or current memory usage, but also "counts" that can go up and down, such as the number of active threads.
+///
+/// Recording a value with a meter:
+/// ```swift
+/// meter.record(100)
+/// ```
 public final class Meter {
-    /// ``_handler`` is only public to allow access from `MetricsTestKit`. Do not consider it part of the public API.
+    /// `_handler` and `_factory` are only public to allow access from `MetricsTestKit`.
+    /// Do not consider them part of the public API.
+    @_documentation(visibility: internal)
     public let _handler: MeterHandler
+    @usableFromInline
+    @_documentation(visibility: internal)
+    package let _factory: MetricsFactory
+    /// The label for the meter.
     public let label: String
+    /// The dimensions for the meter.
     public let dimensions: [(String, String)]
 
-    /// Alternative way to create a new `Meter`, while providing an explicit `MeterHandler`.
+    /// Alternative way to create a new meter, while providing an explicit meter handler.
     ///
-    /// - warning: This initializer provides an escape hatch for situations where one must use a custom factory instead of the global one.
-    ///            We do not expect this API to be used in normal circumstances, so if you find yourself using it make sure it's for a good reason.
-    ///
-    /// - SeeAlso: Use `init(label:dimensions:)` to create `Meter` instances using the configured metrics backend.
+    /// - SeeAlso: Use `init(label:dimensions:)` to create instances of ``Meter`` using the configured metrics backend.
     ///
     /// - parameters:
-    ///     - label: The label for the `Recorder`.
-    ///     - dimensions: The dimensions for the `Recorder`.
-    ///     - handler: The custom backend.
-    public init(label: String, dimensions: [(String, String)], handler: MeterHandler) {
+    ///   - label: The label for the `Recorder`.
+    ///   - dimensions: The dimensions for the `Recorder`.
+    ///   - handler: The custom backend.
+    ///   - factory: The custom metrics factory.
+    public init(label: String, dimensions: [(String, String)], handler: MeterHandler, factory: MetricsFactory) {
         self.label = label
         self.dimensions = dimensions
         self._handler = handler
+        self._factory = factory
     }
 
-    /// Set a value.
+    /// Alternative way to create a new meter, while providing an explicit meter handler.
+    ///
+    /// - SeeAlso: Use `init(label:dimensions:)` to create instances of ``Meter`` using the configured metrics backend.
     ///
     /// - parameters:
-    ///     - value: Value to set.
+    ///   - label: The label for the `Recorder`.
+    ///   - dimensions: The dimensions for the `Recorder`.
+    ///   - handler: The custom backend.
+    public convenience init(label: String, dimensions: [(String, String)], handler: MeterHandler) {
+        self.init(label: label, dimensions: dimensions, handler: handler, factory: MetricsSystem.factory)
+    }
+
+    /// Set an integer value.
+    ///
+    /// - parameters:
+    ///   - value: Value to set.
     @inlinable
     public func set<DataType: BinaryInteger>(_ value: DataType) {
         self._handler.set(Int64(value))
     }
 
-    /// Set a value.
+    /// Set a floating-point value.
     ///
     /// - parameters:
-    ///     - value: Value to est.
+    ///   - value: Value to est.
     @inlinable
     public func set<DataType: BinaryFloatingPoint>(_ value: DataType) {
         self._handler.set(Double(value))
     }
 
-    /// Increment the Meter.
+    /// Increment the meter.
     ///
     /// - parameters:
-    ///     - amount: Amount to increment by.
+    ///   - amount: Amount to increment by.
     @inlinable
     public func increment<DataType: BinaryFloatingPoint>(by amount: DataType) {
         self._handler.increment(by: Double(amount))
     }
 
-    /// Increment the Meter by one.
+    /// Increment the meter by one.
     @inlinable
     public func increment() {
         self.increment(by: 1.0)
     }
 
-    /// Decrement the Meter.
+    /// Decrement the meter.
     ///
     /// - parameters:
-    ///     - amount: Amount to decrement by.
+    ///    - amount: Amount to decrement by.
     @inlinable
     public func decrement<DataType: BinaryFloatingPoint>(by amount: DataType) {
         self._handler.decrement(by: Double(amount))
     }
 
-    /// Decrement the Meter by one.
+    /// Decrement the meter by one.
     @inlinable
     public func decrement() {
         self.decrement(by: 1.0)
@@ -264,21 +404,32 @@ public final class Meter {
 }
 
 extension Meter {
-    /// Create a new `Meter`.
+    /// Create a new meter.
     ///
     /// - parameters:
-    ///     - label: The label for the `Meter`.
-    ///     - dimensions: The dimensions for the `Meter`.
+    ///   - label: The label for the `Meter`.
+    ///   - dimensions: The dimensions for the `Meter`.
+    ///   - factory: The custom metrics factory.
+    public convenience init(label: String, dimensions: [(String, String)] = [], factory: MetricsFactory) {
+        let handler = factory.makeMeter(label: label, dimensions: dimensions)
+        self.init(label: label, dimensions: dimensions, handler: handler, factory: factory)
+    }
+
+    /// Create a new meter.
+    ///
+    /// - parameters:
+    ///   - label: The label for the `Meter`.
+    ///   - dimensions: The dimensions for the `Meter`.
     public convenience init(label: String, dimensions: [(String, String)] = []) {
-        let handler = MetricsSystem.factory.makeMeter(label: label, dimensions: dimensions)
-        self.init(label: label, dimensions: dimensions, handler: handler)
+        self.init(label: label, dimensions: dimensions, factory: MetricsSystem.factory)
     }
 
     /// Signal the underlying metrics library that this recorder will never be updated again.
+    ///
     /// In response the library MAY decide to eagerly release any resources held by this `Recorder`.
     @inlinable
     public func destroy() {
-        MetricsSystem.factory.destroyMeter(self._handler)
+        self._factory.destroyMeter(self._handler)
     }
 }
 
@@ -290,35 +441,75 @@ extension Meter: CustomStringConvertible {
 
 // MARK: - Recorder
 
-/// A recorder collects observations within a time window (usually things like response sizes) and *can* provide aggregated information about the data sample, for example, count, sum, min, max and various quantiles.
+/// A recorder collects observations within a time window.
+///
+/// An example is using a recorder to capture response sizes.
+/// A recorder *can* provide aggregated information about the data sample such as count, sum, min, max, and various quantiles.
 ///
 /// This is the user-facing Recorder API.
+/// Its behavior depends on the ``RecorderHandler`` implementation.
 ///
-/// Its behavior depends on the `RecorderHandler` implementation.
+/// Recording a value:
+///
+/// ```swift
+/// recorder.record(101)
+/// ```
 public class Recorder {
-    /// ``_handler`` is only public to allow access from `MetricsTestKit`. Do not consider it part of the public API.
+    /// `_handler` and `_factory` are only public to allow access from `MetricsTestKit`.
+    /// Do not consider them part of the public API.
+    @_documentation(visibility: internal)
     public let _handler: RecorderHandler
+    @_documentation(visibility: internal)
+    @usableFromInline
+    package let _factory: MetricsFactory
+    /// The label for the recorder.
     public let label: String
+    /// The dimensions for the recorder.
     public let dimensions: [(String, String)]
+    /// A Boolean value that indicates whether to aggregate values.
     public let aggregate: Bool
 
-    /// Alternative way to create a new `Recorder`, while providing an explicit `RecorderHandler`.
-    ///
-    /// - warning: This initializer provides an escape hatch for situations where one must use a custom factory instead of the global one.
-    ///            We do not expect this API to be used in normal circumstances, so if you find yourself using it make sure it's for a good reason.
+    /// Alternative way to create a new recorder, while providing an explicit recorder handler.
     ///
     /// - SeeAlso: Use `init(label:dimensions:)` to create `Recorder` instances using the configured metrics backend.
     ///
     /// - parameters:
-    ///     - label: The label for the `Recorder`.
-    ///     - dimensions: The dimensions for the `Recorder`.
-    ///     - aggregate: aggregate recorded values to produce statistics across a sample size
-    ///     - handler: The custom backend.
-    public init(label: String, dimensions: [(String, String)], aggregate: Bool, handler: RecorderHandler) {
+    ///   - label: The label for the `Recorder`.
+    ///   - dimensions: The dimensions for the `Recorder`.
+    ///   - aggregate: A Boolean value that indicates whether to aggregate values.
+    ///   - handler: The custom backend.
+    ///   - factory: The custom metrics factory.
+    public init(
+        label: String,
+        dimensions: [(String, String)],
+        aggregate: Bool,
+        handler: RecorderHandler,
+        factory: MetricsFactory
+    ) {
         self.label = label
         self.dimensions = dimensions
         self.aggregate = aggregate
         self._handler = handler
+        self._factory = factory
+    }
+
+    /// Alternative way to create a new recorder, while providing an explicit recorder handler.
+    ///
+    /// - SeeAlso: Use `init(label:dimensions:)` to create `Recorder` instances using the configured metrics backend.
+    ///
+    /// - parameters:
+    ///   - label: The label for the `Recorder`.
+    ///   - dimensions: The dimensions for the `Recorder`.
+    ///   - aggregate: A Boolean value that indicates whether to aggregate values.
+    ///   - handler: The custom backend.
+    public convenience init(label: String, dimensions: [(String, String)], aggregate: Bool, handler: RecorderHandler) {
+        self.init(
+            label: label,
+            dimensions: dimensions,
+            aggregate: aggregate,
+            handler: handler,
+            factory: MetricsSystem.factory
+        )
     }
 
     /// Record a value.
@@ -327,7 +518,7 @@ public class Recorder {
     /// This means that the value of this `Recorder` will match the passed in value, rather than accumulate and sum the values up.
     ///
     /// - parameters:
-    ///     - value: Value to record.
+    ///   - value: Value to record.
     @inlinable
     public func record<DataType: BinaryInteger>(_ value: DataType) {
         self._handler.record(Int64(value))
@@ -339,7 +530,7 @@ public class Recorder {
     /// This means that the value of this `Recorder` will match the passed in value, rather than accumulate and sum the values up.
     ///
     /// - parameters:
-    ///     - value: Value to record.
+    ///   - value: Value to record.
     @inlinable
     public func record<DataType: BinaryFloatingPoint>(_ value: DataType) {
         self._handler.record(Double(value))
@@ -347,22 +538,39 @@ public class Recorder {
 }
 
 extension Recorder {
-    /// Create a new `Recorder`.
+    /// Create a new recorder.
     ///
     /// - parameters:
-    ///     - label: The label for the `Recorder`.
-    ///     - dimensions: The dimensions for the `Recorder`.
-    ///     - aggregate: aggregate recorded values to produce statistics across a sample size
+    ///   - label: The label for the `Recorder`.
+    ///   - dimensions: The dimensions for the `Recorder`.
+    ///   - aggregate: A Boolean value that indicates whether to aggregate values.
     public convenience init(label: String, dimensions: [(String, String)] = [], aggregate: Bool = true) {
-        let handler = MetricsSystem.factory.makeRecorder(label: label, dimensions: dimensions, aggregate: aggregate)
-        self.init(label: label, dimensions: dimensions, aggregate: aggregate, handler: handler)
+        self.init(label: label, dimensions: dimensions, aggregate: aggregate, factory: MetricsSystem.factory)
+    }
+
+    /// Create a new recorder using a custom metrics factory that you provide..
+    ///
+    /// - parameters:
+    ///   - label: The label for the `Recorder`.
+    ///   - dimensions: The dimensions for the `Recorder`.
+    ///   - aggregate: A Boolean value that indicates whether to aggregate values.
+    ///   - factory: The custom metrics factory.
+    public convenience init(
+        label: String,
+        dimensions: [(String, String)] = [],
+        aggregate: Bool = true,
+        factory: MetricsFactory
+    ) {
+        let handler = factory.makeRecorder(label: label, dimensions: dimensions, aggregate: aggregate)
+        self.init(label: label, dimensions: dimensions, aggregate: aggregate, handler: handler, factory: factory)
     }
 
     /// Signal the underlying metrics library that this recorder will never be updated again.
+    ///
     /// In response the library MAY decide to eagerly release any resources held by this `Recorder`.
     @inlinable
     public func destroy() {
-        MetricsSystem.factory.destroyRecorder(self._handler)
+        self._factory.destroyRecorder(self._handler)
     }
 }
 
@@ -374,6 +582,7 @@ extension Recorder: CustomStringConvertible {
 
 // MARK: - Timer
 
+/// A unit of time.
 public struct TimeUnit: Equatable, Sendable {
     private enum Code: Equatable, Sendable {
         case nanoseconds
@@ -386,6 +595,7 @@ public struct TimeUnit: Equatable, Sendable {
     }
 
     private let code: Code
+    /// The number of nanoseconds in this time unit.
     public let scaleFromNanoseconds: UInt64
 
     private init(code: Code, scaleFromNanoseconds: UInt64) {
@@ -395,57 +605,90 @@ public struct TimeUnit: Equatable, Sendable {
         self.scaleFromNanoseconds = scaleFromNanoseconds
     }
 
+    /// A nanosecond.
     public static let nanoseconds = TimeUnit(code: .nanoseconds, scaleFromNanoseconds: 1)
+    /// A microsecond.
     public static let microseconds = TimeUnit(code: .microseconds, scaleFromNanoseconds: 1000)
+    /// A millisecond.
     public static let milliseconds = TimeUnit(
         code: .milliseconds,
         scaleFromNanoseconds: 1000 * TimeUnit.microseconds.scaleFromNanoseconds
     )
+    /// A second.
     public static let seconds = TimeUnit(
         code: .seconds,
         scaleFromNanoseconds: 1000 * TimeUnit.milliseconds.scaleFromNanoseconds
     )
+    /// A minute.
     public static let minutes = TimeUnit(
         code: .minutes,
         scaleFromNanoseconds: 60 * TimeUnit.seconds.scaleFromNanoseconds
     )
+    /// An hour.
     public static let hours = TimeUnit(code: .hours, scaleFromNanoseconds: 60 * TimeUnit.minutes.scaleFromNanoseconds)
+    /// A day.
     public static let days = TimeUnit(code: .days, scaleFromNanoseconds: 24 * TimeUnit.hours.scaleFromNanoseconds)
 }
 
-/// A timer collects observations within a time window (usually things like request durations) and provides aggregated information about the data sample,
-/// for example, min, max and various quantiles. It is similar to a `Recorder` but specialized for values that represent durations.
+/// A timer collects observations that represents durations within a time window.
+///
+/// It is similar to a `Recorder` but specialized for values that represent durations, such as request durations.
+/// A time provides  aggregated information about the data sample, such as min, max, and various quantiles.
 ///
 /// This is the user-facing Timer API.
+/// Its behavior depends on the ``TimerHandler`` implementation.
 ///
-/// Its behavior depends on the `TimerHandler` implementation.
+/// Recording a value:
+///
+/// ```swift
+/// timer.recordMilliseconds(52)
+/// ```
+
 public final class Timer {
-    /// ``_handler`` is only public to allow access from `MetricsTestKit`. Do not consider it part of the public API.
+    /// `_handler` and `_factory` are only public to allow access from `MetricsTestKit`.
+    /// Do not consider them part of the public API.
+    @_documentation(visibility: internal)
     public let _handler: TimerHandler
+    @_documentation(visibility: internal)
+    @usableFromInline
+    package let _factory: MetricsFactory
+    /// The label for the timer.
     public let label: String
+    /// The dimensions for the timer.
     public let dimensions: [(String, String)]
 
-    /// Alternative way to create a new `Timer`, while providing an explicit `TimerHandler`.
+    /// Alternative way to create a new timer, while providing an explicit timer handler.
     ///
-    /// - warning: This initializer provides an escape hatch for situations where one must use a custom factory instead of the global one.
-    ///            We do not expect this API to be used in normal circumstances, so if you find yourself using it make sure it's for a good reason.
-    ///
-    /// - SeeAlso: Use `init(label:dimensions:)` to create `Recorder` instances using the configured metrics backend.
+    /// - SeeAlso: Use `init(label:dimensions:)` to create instances of ``Timer`` using the configured metrics backend.
     ///
     /// - parameters:
-    ///     - label: The label for the `Timer`.
-    ///     - dimensions: The dimensions for the `Timer`.
-    ///     - handler: The custom backend.
-    public init(label: String, dimensions: [(String, String)], handler: TimerHandler) {
+    ///   - label: The label for the `Timer`.
+    ///   - dimensions: The dimensions for the `Timer`.
+    ///   - handler: The custom backend.
+    ///   - factory: The custom factory.
+    public init(label: String, dimensions: [(String, String)], handler: TimerHandler, factory: MetricsFactory) {
         self.label = label
         self.dimensions = dimensions
         self._handler = handler
+        self._factory = factory
+    }
+
+    /// Alternative way to create a new timer, while providing an explicit timer handler.
+    ///
+    /// - SeeAlso: Use `init(label:dimensions:)` to create instances of ``Timer`` using the configured metrics backend.
+    ///
+    /// - parameters:
+    ///   - label: The label for the `Timer`.
+    ///   - dimensions: The dimensions for the `Timer`.
+    ///   - handler: The custom backend.
+    public convenience init(label: String, dimensions: [(String, String)], handler: TimerHandler) {
+        self.init(label: label, dimensions: dimensions, handler: handler, factory: MetricsSystem.factory)
     }
 
     /// Record a duration in nanoseconds.
     ///
     /// - parameters:
-    ///     - duration: Duration to record.
+    ///   - duration: Duration to record.
     @inlinable
     public func recordNanoseconds(_ duration: Int64) {
         self._handler.recordNanoseconds(duration)
@@ -454,7 +697,7 @@ public final class Timer {
     /// Record a duration in nanoseconds.
     ///
     /// - parameters:
-    ///     - duration: Duration to record.
+    ///   - duration: Duration to record.
     @inlinable
     public func recordNanoseconds<DataType: BinaryInteger>(_ duration: DataType) {
         self.recordNanoseconds(duration >= Int64.max ? Int64.max : Int64(duration))
@@ -463,7 +706,7 @@ public final class Timer {
     /// Record a duration in microseconds.
     ///
     /// - parameters:
-    ///     - duration: Duration to record.
+    ///   - duration: Duration to record.
     @inlinable
     public func recordMicroseconds<DataType: BinaryInteger>(_ duration: DataType) {
         guard duration <= Int64.max else { return self.recordNanoseconds(Int64.max) }
@@ -479,7 +722,7 @@ public final class Timer {
     /// Record a duration in microseconds.
     ///
     /// - parameters:
-    ///     - duration: Duration to record.
+    ///   - duration: Duration to record.
     @inlinable
     public func recordMicroseconds<DataType: BinaryFloatingPoint>(_ duration: DataType) {
         self.recordNanoseconds(Double(duration * 1000) < Double(Int64.max) ? Int64(duration * 1000) : Int64.max)
@@ -488,7 +731,7 @@ public final class Timer {
     /// Record a duration in milliseconds.
     ///
     /// - parameters:
-    ///     - duration: Duration to record.
+    ///   - duration: Duration to record.
     @inlinable
     public func recordMilliseconds<DataType: BinaryInteger>(_ duration: DataType) {
         guard duration <= Int64.max else { return self.recordNanoseconds(Int64.max) }
@@ -504,7 +747,7 @@ public final class Timer {
     /// Record a duration in milliseconds.
     ///
     /// - parameters:
-    ///     - duration: Duration to record.
+    ///   - duration: Duration to record.
     @inlinable
     public func recordMilliseconds<DataType: BinaryFloatingPoint>(_ duration: DataType) {
         self.recordNanoseconds(
@@ -515,7 +758,7 @@ public final class Timer {
     /// Record a duration in seconds.
     ///
     /// - parameters:
-    ///     - duration: Duration to record.
+    ///   - duration: Duration to record.
     @inlinable
     public func recordSeconds<DataType: BinaryInteger>(_ duration: DataType) {
         guard duration <= Int64.max else { return self.recordNanoseconds(Int64.max) }
@@ -531,7 +774,7 @@ public final class Timer {
     /// Record a duration in seconds.
     ///
     /// - parameters:
-    ///     - duration: Duration to record.
+    ///   - duration: Duration to record.
     @inlinable
     public func recordSeconds<DataType: BinaryFloatingPoint>(_ duration: DataType) {
         self.recordNanoseconds(
@@ -541,37 +784,69 @@ public final class Timer {
 }
 
 extension Timer {
-    /// Create a new `Timer`.
+    /// Create a new timer.
     ///
     /// - parameters:
-    ///     - label: The label for the `Timer`.
-    ///     - dimensions: The dimensions for the `Timer`.
-    public convenience init(label: String, dimensions: [(String, String)] = []) {
-        let handler = MetricsSystem.factory.makeTimer(label: label, dimensions: dimensions)
-        self.init(label: label, dimensions: dimensions, handler: handler)
+    ///   - label: The label for the `Timer`.
+    ///   - dimensions: The dimensions for the `Timer`.
+    ///   - factory: The custom factory.
+    public convenience init(label: String, dimensions: [(String, String)] = [], factory: MetricsFactory) {
+        let handler = factory.makeTimer(label: label, dimensions: dimensions)
+        self.init(label: label, dimensions: dimensions, handler: handler, factory: factory)
     }
 
-    /// Create a new `Timer`.
+    /// Create a new timer using a custom metrics factory that you provide..
     ///
     /// - parameters:
-    ///     - label: The label for the `Timer`.
-    ///     - dimensions: The dimensions for the `Timer`.
-    ///     - displayUnit: A hint to the backend responsible for presenting the data of the preferred display unit. This is not guaranteed to be supported by all backends.
+    ///   - label: The label for the `Timer`.
+    ///   - dimensions: The dimensions for the `Timer`.
+    public convenience init(label: String, dimensions: [(String, String)] = []) {
+        self.init(label: label, dimensions: dimensions, factory: MetricsSystem.factory)
+    }
+
+    /// Create a new timer.
+    ///
+    /// - parameters:
+    ///   - label: The label for the `Timer`.
+    ///   - dimensions: The dimensions for the `Timer`.
+    ///   - displayUnit: A hint to the backend responsible for presenting the data of the preferred display unit. This is not guaranteed to be supported by all backends.
+    ///   - factory: The custom factory.
+    public convenience init(
+        label: String,
+        dimensions: [(String, String)] = [],
+        preferredDisplayUnit displayUnit: TimeUnit,
+        factory: MetricsFactory
+    ) {
+        let handler = factory.makeTimer(label: label, dimensions: dimensions)
+        handler.preferDisplayUnit(displayUnit)
+        self.init(label: label, dimensions: dimensions, handler: handler, factory: factory)
+    }
+
+    /// Create a new timer.
+    ///
+    /// - parameters:
+    ///   - label: The label for the `Timer`.
+    ///   - dimensions: The dimensions for the `Timer`.
+    ///   - displayUnit: A hint to the backend responsible for presenting the data of the preferred display unit. This is not guaranteed to be supported by all backends.
     public convenience init(
         label: String,
         dimensions: [(String, String)] = [],
         preferredDisplayUnit displayUnit: TimeUnit
     ) {
-        let handler = MetricsSystem.factory.makeTimer(label: label, dimensions: dimensions)
-        handler.preferDisplayUnit(displayUnit)
-        self.init(label: label, dimensions: dimensions, handler: handler)
+        self.init(
+            label: label,
+            dimensions: dimensions,
+            preferredDisplayUnit: displayUnit,
+            factory: MetricsSystem.factory
+        )
     }
 
     /// Signal the underlying metrics library that this timer will never be updated again.
+    ///
     /// In response the library MAY decide to eagerly release any resources held by this `Timer`.
     @inlinable
     public func destroy() {
-        MetricsSystem.factory.destroyTimer(self._handler)
+        self._factory.destroyTimer(self._handler)
     }
 }
 
@@ -583,18 +858,21 @@ extension Timer: CustomStringConvertible {
 
 // MARK: - MetricsSystem
 
-/// The `MetricsSystem` is a global facility where the default metrics backend implementation (`MetricsFactory`) can be
-/// configured. `MetricsSystem` is set up just once in a given program to set up the desired metrics backend
-/// implementation.
+/// A global facility where the default metrics backend implementation is configured.
+///
+/// `MetricsSystem` is set up just once in a given program to create the desired metrics backend
+/// implementation using ``MetricsFactory``.
 public enum MetricsSystem {
     private static let _factory = FactoryBox(NOOPMetricsHandler.instance)
 
-    /// `bootstrap` is an one-time configuration function which globally selects the desired metrics backend
-    /// implementation. `bootstrap` can be called at maximum once in any given program, calling it more than once will
+    /// A one-time configuration function which globally selects the desired metrics backend
+    /// implementation.
+    ///
+    /// `bootstrap` can be called at maximum once in any given program, calling it more than once will
     /// lead to undefined behavior, most likely a crash.
     ///
     /// - parameters:
-    ///     - factory: A factory that given an identifier produces instances of metrics handlers such as `CounterHandler`, `RecorderHandler` and `TimerHandler`.
+    ///   - factory: A factory that given an identifier produces instances of metrics handlers such as ``CounterHandler``, ``RecorderHandler``, or  ``TimerHandler``.
     public static func bootstrap(_ factory: MetricsFactory) {
         self._factory.replaceFactory(factory, validate: true)
     }
@@ -655,22 +933,22 @@ public enum MetricsSystem {
 // MARK: - MetricsFactory
 
 /// The `MetricsFactory` is the bridge between the `MetricsSystem` and the metrics backend implementation.
-/// `MetricsFactory`'s role is to initialize concrete implementations of the various metric types:
+///
+/// The role of `MetricsFactory` is to initialize concrete implementations of the various metric types:
 /// * `Counter` -> `CounterHandler`
 /// * `FloatingPointCounter` -> `FloatingPointCounterHandler`
 /// * `Recorder` -> `RecorderHandler`
 /// * `Timer` -> `TimerHandler`
 ///
-/// - warning: This type is an implementation detail and should not be used directly, unless implementing your own metrics backend.
-///            To use the SwiftMetrics API, please refer to the documentation of `MetricsSystem`.
+/// To use the SwiftMetrics API, please refer to the documentation of `MetricsSystem`.
 ///
-/// # Destroying metrics
+/// ### Destroying metrics
 ///
 /// Since _some_ metrics implementations may need to allocate (potentially "heavy") resources for metrics, destroying
 /// metrics offers a signal to libraries when a metric is "known to never be updated again."
 ///
 /// While many metrics are bound to the entire lifetime of an application and thus never need to be destroyed eagerly,
-/// some metrics have well defined unique life-cycles, and it may be beneficial to release any resources held by them
+/// some metrics have well defined unique life-cycles where it may be beneficial to release any resources held by them
 /// more eagerly than awaiting the application's termination. In such cases, a library or application should invoke
 /// a metric's appropriate `destroy()` method, which in turn results in the corresponding handler that it is backed by
 /// to be passed to `destroyCounter(handler:)`, `destroyRecorder(handler:)` or `destroyTimer(handler:)` where the factory
@@ -680,75 +958,80 @@ public enum MetricsSystem {
 /// libraries using the metrics API should always assume a library WILL make use of this signal, and shall not
 /// neglect calling these methods when appropriate.
 public protocol MetricsFactory: _SwiftMetricsSendableProtocol {
-    /// Create a backing `CounterHandler`.
+    /// Create a backing counter handler.
     ///
     /// - parameters:
-    ///     - label: The label for the `CounterHandler`.
-    ///     - dimensions: The dimensions for the `CounterHandler`.
+    ///   - label: The label for the `CounterHandler`.
+    ///   - dimensions: The dimensions for the `CounterHandler`.
     func makeCounter(label: String, dimensions: [(String, String)]) -> CounterHandler
 
-    /// Create a backing `FloatingPointCounterHandler`.
+    /// Create a backing floating-point handler.
     ///
     /// - parameters:
-    ///     - label: The label for the `FloatingPointCounterHandler`.
-    ///     - dimensions: The dimensions for the `FloatingPointCounterHandler`.
+    ///   - label: The label for the `FloatingPointCounterHandler`.
+    ///   - dimensions: The dimensions for the `FloatingPointCounterHandler`.
     func makeFloatingPointCounter(label: String, dimensions: [(String, String)]) -> FloatingPointCounterHandler
 
-    /// Create a backing `MeterHandler`.
+    /// Create a backing meter handler.
     ///
     /// - parameters:
-    ///     - label: The label for the `MeterHandler`.
-    ///     - dimensions: The dimensions for the `MeterHandler`.
+    ///   - label: The label for the `MeterHandler`.
+    ///   - dimensions: The dimensions for the `MeterHandler`.
     func makeMeter(label: String, dimensions: [(String, String)]) -> MeterHandler
 
-    /// Create a backing `RecorderHandler`.
+    /// Create a backing recorder handler.
     ///
     /// - parameters:
-    ///     - label: The label for the `RecorderHandler`.
-    ///     - dimensions: The dimensions for the `RecorderHandler`.
-    ///     - aggregate: Is data aggregation expected.
+    ///   - label: The label for the `RecorderHandler`.
+    ///   - dimensions: The dimensions for the `RecorderHandler`.
+    ///   - aggregate: A Boolean value that indicates whether to aggregate values.
     func makeRecorder(label: String, dimensions: [(String, String)], aggregate: Bool) -> RecorderHandler
 
-    /// Create a backing `TimerHandler`.
+    /// Create a backing timer handler.
     ///
     /// - parameters:
-    ///     - label: The label for the `TimerHandler`.
-    ///     - dimensions: The dimensions for the `TimerHandler`.
+    ///   - label: The label for the `TimerHandler`.
+    ///   - dimensions: The dimensions for the `TimerHandler`.
     func makeTimer(label: String, dimensions: [(String, String)]) -> TimerHandler
 
-    /// Invoked when the corresponding `Counter`'s `destroy()` function is invoked.
+    /// Invoked when the corresponding counter's `destroy()` function is invoked.
+    ///
     /// Upon receiving this signal the factory may eagerly release any resources related to this counter.
     ///
     /// - parameters:
-    ///     - handler: The handler to be destroyed.
+    ///   - handler: The handler to be destroyed.
     func destroyCounter(_ handler: CounterHandler)
 
-    /// Invoked when the corresponding `Meter`'s `destroy()` function is invoked.
+    /// Invoked when the corresponding meter's `destroy()` function is invoked.
+    ///
     /// Upon receiving this signal the factory may eagerly release any resources related to this recorder.
     ///
     /// - parameters:
-    ///     - handler: The handler to be destroyed.
+    ///   - handler: The handler to be destroyed.
     func destroyMeter(_ handler: MeterHandler)
 
-    /// Invoked when the corresponding `FloatingPointCounter`'s `destroy()` function is invoked.
+    /// Invoked when the corresponding floating-point counter's `destroy()` function is invoked.
+    ///
     /// Upon receiving this signal the factory may eagerly release any resources related to this counter.
     ///
     /// - parameters:
-    ///     - handler: The handler to be destroyed.
+    ///   - handler: The handler to be destroyed.
     func destroyFloatingPointCounter(_ handler: FloatingPointCounterHandler)
 
-    /// Invoked when the corresponding `Recorder`'s `destroy()` function is invoked.
+    /// Invoked when the corresponding recorder's `destroy()` function is invoked.
+    ///
     /// Upon receiving this signal the factory may eagerly release any resources related to this recorder.
     ///
     /// - parameters:
-    ///     - handler: The handler to be destroyed.
+    ///   - handler: The handler to be destroyed.
     func destroyRecorder(_ handler: RecorderHandler)
 
-    /// Invoked when the corresponding `Timer`'s `destroy()` function is invoked.
+    /// Invoked when the corresponding Timer's `destroy()` function is invoked.
+    ///
     /// Upon receiving this signal the factory may eagerly release any resources related to this timer.
     ///
     /// - parameters:
-    ///     - handler: The handler to be destroyed.
+    ///   - handler: The handler to be destroyed.
     func destroyTimer(_ handler: TimerHandler)
 }
 
@@ -756,11 +1039,12 @@ public protocol MetricsFactory: _SwiftMetricsSendableProtocol {
 internal final class AccumulatingRoundingFloatingPointCounter: FloatingPointCounterHandler {
     private let lock = Lock()
     private let counterHandler: CounterHandler
+    private let factory: MetricsFactory
     internal var fraction: Double = 0
 
-    init(label: String, dimensions: [(String, String)]) {
-        self.counterHandler = MetricsSystem
-            .factory.makeCounter(label: label, dimensions: dimensions)
+    init(label: String, dimensions: [(String, String)], factory: MetricsFactory) {
+        self.counterHandler = factory.makeCounter(label: label, dimensions: dimensions)
+        self.factory = factory
     }
 
     func increment(by amount: Double) {
@@ -813,7 +1097,7 @@ internal final class AccumulatingRoundingFloatingPointCounter: FloatingPointCoun
     }
 
     func destroy() {
-        MetricsSystem.factory.destroyCounter(self.counterHandler)
+        self.factory.destroyCounter(self.counterHandler)
     }
 }
 
@@ -824,10 +1108,11 @@ internal final class AccumulatingMeter: MeterHandler, @unchecked Sendable {
     // FIXME: use swift-atomics when floating point support is available
     private var value: Double = 0
     private let lock = Lock()
+    private let factory: MetricsFactory
 
-    init(label: String, dimensions: [(String, String)]) {
-        self.recorderHandler = MetricsSystem
-            .factory.makeRecorder(label: label, dimensions: dimensions, aggregate: true)
+    init(label: String, dimensions: [(String, String)], factory: MetricsFactory) {
+        self.recorderHandler = factory.makeRecorder(label: label, dimensions: dimensions, aggregate: true)
+        self.factory = factory
     }
 
     func set(_ value: Int64) {
@@ -898,53 +1183,55 @@ internal final class AccumulatingMeter: MeterHandler, @unchecked Sendable {
     }
 
     func destroy() {
-        MetricsSystem.factory.destroyRecorder(self.recorderHandler)
+        self.factory.destroyRecorder(self.recorderHandler)
     }
 }
 
 extension MetricsFactory {
-    /// Create a default backing `FloatingPointCounterHandler` for backends which do not naively support floating point counters.
+    /// Create a default backing floating-point counter handler for backends which do not naively support floating point counters.
     ///
-    /// The created FloatingPointCounterHandler is a wrapper around a backend's CounterHandler which accumulates floating point values and records increments to an underlying CounterHandler after crossing integer boundaries.
+    /// The created floating-point counter handler is a wrapper around a backend's counter handler which accumulates floating point values and records increments to an underlying counter handler after crossing integer boundaries.
     ///
     /// - parameters:
-    ///     - label: The label for the `FloatingPointCounterHandler`.
-    ///     - dimensions: The dimensions for the `FloatingPointCounterHandler`.
+    ///   - label: The label for the `FloatingPointCounterHandler`.
+    ///   - dimensions: The dimensions for the `FloatingPointCounterHandler`.
     public func makeFloatingPointCounter(label: String, dimensions: [(String, String)]) -> FloatingPointCounterHandler {
-        AccumulatingRoundingFloatingPointCounter(label: label, dimensions: dimensions)
+        AccumulatingRoundingFloatingPointCounter(label: label, dimensions: dimensions, factory: self)
     }
 
-    /// Invoked when the corresponding `FloatingPointCounter`'s `destroy()` function is invoked.
+    /// Invoked when the corresponding floating-point counter's `destroy()` function is invoked.
+    ///
     /// Upon receiving this signal the factory may eagerly release any resources related to this counter.
     ///
     /// `destroyFloatingPointCounter` must be implemented if `makeFloatingPointCounter` is implemented.
     ///
     /// - parameters:
-    ///     - handler: The handler to be destroyed.
+    ///   - handler: The handler to be destroyed.
     public func destroyFloatingPointCounter(_ handler: FloatingPointCounterHandler) {
         (handler as? AccumulatingRoundingFloatingPointCounter)?.destroy()
     }
 }
 
 extension MetricsFactory {
-    /// Create a default backing `MeterHandler` for backends which do not naively support meters.
+    /// Create a default backing meter handler for backends which do not naively support meters.
     ///
     /// The created MeterHandler is a wrapper around a backend's RecorderHandler which records current values.
     ///
     /// - parameters:
-    ///     - label: The label for the `MeterHandler`.
-    ///     - dimensions: The dimensions for the `MeterHandler`.
+    ///   - label: The label for the `MeterHandler`.
+    ///   - dimensions: The dimensions for the `MeterHandler`.
     public func makeMeter(label: String, dimensions: [(String, String)]) -> MeterHandler {
-        AccumulatingMeter(label: label, dimensions: dimensions)
+        AccumulatingMeter(label: label, dimensions: dimensions, factory: self)
     }
 
-    /// Invoked when the corresponding `Meter`'s `destroy()` function is invoked.
+    /// Invoked when the corresponding meter's `destroy()` function is invoked.
+    ///
     /// Upon receiving this signal the factory may eagerly release any resources related to this counter.
     ///
     /// `destroyMeter` must be implemented if `makeMeter` is implemented.
     ///
     /// - parameters:
-    ///     - handler: The handler to be destroyed.
+    ///   - handler: The handler to be destroyed.
     public func destroyMeter(_ handler: MeterHandler) {
         (handler as? AccumulatingMeter)?.destroy()
     }
@@ -952,12 +1239,12 @@ extension MetricsFactory {
 
 // MARK: - Backend Handlers
 
-/// A `CounterHandler` represents a backend implementation of a `Counter`.
+/// A counter handler represents a backend implementation of a counter.
 ///
 /// This type is an implementation detail and should not be used directly, unless implementing your own metrics backend.
 /// To use the SwiftMetrics API, please refer to the documentation of `Counter`.
 ///
-/// # Implementation requirements
+/// ### Implementation requirements
 ///
 /// To implement your own `CounterHandler` you should respect a few requirements that are necessary so applications work
 /// as expected regardless of the selected `CounterHandler` implementation.
@@ -967,19 +1254,19 @@ public protocol CounterHandler: AnyObject, _SwiftMetricsSendableProtocol {
     /// Increment the counter.
     ///
     /// - parameters:
-    ///     - by: Amount to increment by.
+    ///   - by: Amount to increment by.
     func increment(by: Int64)
 
     /// Reset the counter back to zero.
     func reset()
 }
 
-/// A `FloatingPointCounterHandler` represents a backend implementation of a `FloatingPointCounter`.
+/// A floating-point counter handler represents a backend implementation of a floating-point counter.
 ///
 /// This type is an implementation detail and should not be used directly, unless implementing your own metrics backend.
 /// To use the SwiftMetrics API, please refer to the documentation of `FloatingPointCounter`.
 ///
-/// # Implementation requirements
+/// ### Implementation requirements
 ///
 /// To implement your own `FloatingPointCounterHandler` you should respect a few requirements that are necessary so applications work
 /// as expected regardless of the selected `FloatingPointCounterHandler` implementation.
@@ -989,100 +1276,103 @@ public protocol FloatingPointCounterHandler: AnyObject, _SwiftMetricsSendablePro
     /// Increment the counter.
     ///
     /// - parameters:
-    ///     - by: Amount to increment by.
+    ///   - by: Amount to increment by.
     func increment(by: Double)
 
     /// Reset the counter back to zero.
     func reset()
 }
 
-/// A `RecorderHandler` represents a backend implementation of a `Recorder`.
+/// A recorder handler represents a backend implementation of a recorder.
 ///
 /// This type is an implementation detail and should not be used directly, unless implementing your own metrics backend.
 /// To use the SwiftMetrics API, please refer to the documentation of `Recorder`.
 ///
-/// # Implementation requirements
+/// ### Implementation requirements
 ///
 /// To implement your own `RecorderHandler` you should respect a few requirements that are necessary so applications work
 /// as expected regardless of the selected `RecorderHandler` implementation.
 ///
 /// - The `RecorderHandler` must be a `class`.
 public protocol RecorderHandler: AnyObject, _SwiftMetricsSendableProtocol {
-    /// Record a value.
+    /// Record an integer value.
     ///
     /// - parameters:
-    ///     - value: Value to record.
+    ///   - value: Value to record.
     func record(_ value: Int64)
-    /// Record a value.
+    /// Record a floating-point value.
     ///
     /// - parameters:
-    ///     - value: Value to record.
+    ///   - value: Value to record.
     func record(_ value: Double)
 }
 
-/// A `MeterHandler` represents a backend implementation of a `Meter`.
+/// A meter handler represents a backend implementation of a meter.
 ///
 /// This type is an implementation detail and should not be used directly, unless implementing your own metrics backend.
 /// To use the SwiftMetrics API, please refer to the documentation of `Meter`.
 ///
-/// # Implementation requirements
+/// ### Implementation requirements
 ///
 /// To implement your own `MeterHandler` you should respect a few requirements that are necessary so applications work
 /// as expected regardless of the selected `MeterHandler` implementation.
 ///
 /// - The `RecorderHandler` must be a `class`.
 public protocol MeterHandler: AnyObject, _SwiftMetricsSendableProtocol {
-    /// Set a value.
+    /// Set an integer value.
     ///
     /// - parameters:
-    ///     - value: Value to set.
+    ///   - value: Value to set.
     func set(_ value: Int64)
-    /// Set a value.
+    /// Set a floating-point value.
     ///
     /// - parameters:
-    ///     - value: Value to set.
+    ///   - value: Value to set.
     func set(_ value: Double)
 
     /// Increment the value.
     ///
     /// - parameters:
-    ///     - by: Amount to increment by.
+    ///   - by: Amount to increment by.
     func increment(by: Double)
 
     /// Decrement the value.
     ///
     /// - parameters:
-    ///     - by: Amount to increment by.
+    ///   - by: Amount to increment by.
     func decrement(by: Double)
 }
 
-/// A `TimerHandler` represents a backend implementation of a `Timer`.
+/// A timer handler represents a backend implementation of a timer.
 ///
 /// This type is an implementation detail and should not be used directly, unless implementing your own metrics backend.
 /// To use the SwiftMetrics API, please refer to the documentation of `Timer`.
 ///
-/// # Implementation requirements
+/// ### Implementation requirements
 ///
 /// To implement your own `TimerHandler` you should respect a few requirements that are necessary so applications work
 /// as expected regardless of the selected `TimerHandler` implementation.
 ///
 /// - The `TimerHandler` must be a `class`.
 public protocol TimerHandler: AnyObject, _SwiftMetricsSendableProtocol {
+
     /// Record a duration in nanoseconds.
     ///
     /// - parameters:
-    ///     - duration: Duration to record.
+    ///   - duration: Duration to record.
     func recordNanoseconds(_ duration: Int64)
 
-    /// Set the preferred display unit for this TimerHandler.
+    /// Set the preferred display unit for this timer handler.
     ///
     /// - parameters:
-    ///     - unit: A hint to the backend responsible for presenting the data of the preferred display unit. This is not guaranteed to be supported by all backends.
+    ///   - unit: A hint to the backend responsible for presenting the data of the preferred display unit. This is not guaranteed to be supported by all backends.
     func preferDisplayUnit(_ unit: TimeUnit)
 }
 
 extension TimerHandler {
-    public func preferDisplayUnit(_: TimeUnit) {
+    /// Set the preferred display unit for this timer handler.
+    /// - Parameter unit: A hint to the backend responsible for presenting the data of the preferred display unit. This is not guaranteed to be supported by all backends.
+    public func preferDisplayUnit(_ unit: TimeUnit) {
         // NOOP
     }
 }
@@ -1092,54 +1382,87 @@ extension TimerHandler {
 /// A pseudo-metrics handler that can be used to send messages to multiple other metrics handlers.
 public final class MultiplexMetricsHandler: MetricsFactory {
     private let factories: [MetricsFactory]
+    /// Creates a new multiplex metrics handler from the metric factories you provide.
+    /// - Parameter factories: The metric factories to multiplex together.
     public init(factories: [MetricsFactory]) {
         self.factories = factories
     }
 
+    /// Creates a new counter handler.
+    /// - Parameters:
+    ///   - label: The label for the `CounterHandler`.
+    ///   - dimensions: The dimensions for the `CounterHandler`.
     public func makeCounter(label: String, dimensions: [(String, String)]) -> CounterHandler {
         MuxCounter(factories: self.factories, label: label, dimensions: dimensions)
     }
 
+    /// Creates a new floating point counter handler.
+    /// - Parameters:
+    ///   - label: The label for the `FloatingPointCounterHandler`.
+    ///   - dimensions: The dimensions for the `FloatingPointCounterHandler`.
     public func makeFloatingPointCounter(label: String, dimensions: [(String, String)]) -> FloatingPointCounterHandler {
         MuxFloatingPointCounter(factories: self.factories, label: label, dimensions: dimensions)
     }
 
+    /// Creates a new meter handler.
+    /// - Parameters:
+    ///   - label: The label for the `MeterHandler`.
+    ///   - dimensions: The dimensions for the `MeterHandler`.
     public func makeMeter(label: String, dimensions: [(String, String)]) -> MeterHandler {
         MuxMeter(factories: self.factories, label: label, dimensions: dimensions)
     }
 
+    /// Creates a new recorder handler.
+    /// - Parameters:
+    ///   - label: The label for the `RecorderHandler`.
+    ///   - dimensions: The dimensions for the `RecorderHandler`.
+    ///   - aggregate: A Boolean value that indicates whether to aggregate values..
     public func makeRecorder(label: String, dimensions: [(String, String)], aggregate: Bool) -> RecorderHandler {
         MuxRecorder(factories: self.factories, label: label, dimensions: dimensions, aggregate: aggregate)
     }
 
+    /// Creates a new timer handler.
+    /// - Parameters:
+    ///   - label: The label for the `TimerHandler`.
+    ///   - dimensions: The dimensions for the `TimerHandler`.
     public func makeTimer(label: String, dimensions: [(String, String)]) -> TimerHandler {
         MuxTimer(factories: self.factories, label: label, dimensions: dimensions)
     }
 
+    /// Signal the underlying metrics library that this counter will never be updated again.
+    /// - Parameter handler: The counter handler to signal.
     public func destroyCounter(_ handler: CounterHandler) {
         for factory in self.factories {
             factory.destroyCounter(handler)
         }
     }
 
+    /// Signal the underlying metrics library that this floating point counter will never be updated again.
+    /// - Parameter handler: The floating point counter handler to signal.
     public func destroyFloatingPointCounter(_ handler: FloatingPointCounterHandler) {
         for factory in self.factories {
             factory.destroyFloatingPointCounter(handler)
         }
     }
 
+    /// Signal the underlying metrics library that this meter will never be updated again.
+    /// - Parameter handler: The meter handler to signal.
     public func destroyMeter(_ handler: MeterHandler) {
         for factory in self.factories {
             factory.destroyMeter(handler)
         }
     }
 
+    /// Signal the underlying metrics library that this recorder will never be updated again.
+    /// - Parameter handler: The recorder handler to signal.
     public func destroyRecorder(_ handler: RecorderHandler) {
         for factory in self.factories {
             factory.destroyRecorder(handler)
         }
     }
 
+    /// Signal the underlying metrics library that this timer will never be updated again.
+    /// - Parameter handler: The timer handler to signal.
     public func destroyTimer(_ handler: TimerHandler) {
         for factory in self.factories {
             factory.destroyTimer(handler)
@@ -1232,10 +1555,13 @@ public final class MultiplexMetricsHandler: MetricsFactory {
     }
 }
 
-/// Ships with the metrics module, used for initial bootstrapping.
+/// A metrics handler that implements the protocols but does nothing.
+///
+/// The no-op metrics handler ships with the metrics module, and is used by default unless a different metrics backend is bootstrapped.
 public final class NOOPMetricsHandler: MetricsFactory, CounterHandler, FloatingPointCounterHandler, MeterHandler,
     RecorderHandler, TimerHandler
 {
+    /// A sharable instance of a No-op metrics handler.
     public static let instance = NOOPMetricsHandler()
 
     private init() {}
